spi: sun6i: add quirk for in-controller clock divider
Previously SPI controllers in Allwinner SoCs has a clock divider inside.
However now the clock divider is removed and to set the transfer clock
rate it's only needed to set the SPI module clock to the target value.

Add a quirk for this kind of SPI controllers.

As the current device used to test this patch lacks a MISO pin (a
write-only MIPI DBI Type-C panel), setting of sampling mode is left out
as a TODO.

Signed-off-by: Icenowy Zheng <icenowy@sipeed.com>