spi: sun6i: add support for R329 SPI controllers
R329 has two SPI controllers. One of it is quite similar to previous
ones, but with internal clock divider removed; the other added MIPI DBI
Type-C offload based on the first one.

Add basical support for these controllers. As the functionality of the
DBI-offload-capable controller is just a superset of the other one, only
add support for the regular one is necessary.

Signed-off-by: Icenowy Zheng <icenowy@sipeed.com>